Add solution to day 17
Started with an A* attempt but it did not choose the right path. Moved
to dijkstra but did not figure out the right key for checked nodes.
Forgot the amount of straight lines made as the key and got wrong
answers.

When I did test it out, it was too slow with the real input. After some
investigation, I figured my problem was not using a priority queue.

Took a solution from online and it worked. First time having to use
external modules.