Add solution to day 20 part 2
After a long time of trying to brute force, optimize and trying to cache
something. I was lost.

Had to get inspired a bit, and it turns out it was one of those
days focusing on "analyze the input". A lack of an example for guiding
the solution also did not help.